+++ main.go
@@ -27,6 +27,7 @@ func main() {
 			Assets: assets,
 		},
 		OnStartup: app.startup,
+		OnShutdown: app.shutdown,
 		Bind: []interface{}{
 			app,
 		},